create TailHelper
